requirements.txt
appnope==0.1.0
backcall==0.1.0
boto3==1.9.86
botocore==1.12.86
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==2.1.5
django-storages==1.7.1
docutils==0.14
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jmespath==0.9.3
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
